requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
joblib
PyQt6
Pillow
torch
torchaudio
skl2onnx
onnx
pytest
